+++ lib/rabl/engine.rb
@@ -62,7 +62,7 @@ module Rabl
           template = @_options[:template] || @virtual_path
 
           digest =
-            if Rails.version.to_s =~ /^[67]/
+            if Rails.version.to_s =~ /^[678]/
               Digestor.digest(name: template, finder: lookup_context, format: :rabl)
             elsif Gem::Version.new(Rails.version) >= Gem::Version.new('4.1')
               Digestor.digest(:name => template, :finder => lookup_context)
@@ -399,7 +399,7 @@ module Rabl
       end
 
       def digestor_available?
-        defined?(Rails) && Rails.version =~ /^[4567]/
+        defined?(Rails) && Rails.version =~ /^[45678]/
       end
 
       def valid_format?(format)